feat: added complete bet settlement ecosystem with automated payout processing

+++ backend/pracbet/src/main/java/com/pracbet/pracbet/ExportAPI/dtos/OddsDto.java
@@ -1,8 +1,10 @@
 package com.pracbet.pracbet.ExportAPI.dtos;
 
+import java.math.BigDecimal;
+
 public record OddsDto(
         String betType,
         String value,
-        Double odd
+        BigDecimal odd
 ) {
 }
